feat: extend default settings in AsyncVectorStore (#2602)
* feat: extend default settings in AsyncVectorStore

This commit introduces a new parameter, custom_settings, to the AsyncVectorStore class. This allows users to provide their own settings that will extend the default settings. This increases the flexibility of the class and allows it to be tailored to specific use cases. The custom settings are applied in the _create_index_if_not_exists method.

* Update vectorstore.py

* Update vectorstore.py

apply changes in vectorstore

* Update vectorstore.py

format the py file

* Update test_vectorstore.py

add test_custom_index_settings in test_vectorstore

* Update test_vectorstore.py

* Update vectorstore.py

fix file format

* Update test_vectorstore.py

fix format

* Update vectorstore.py

add error tips in vectorstore  when confilicting the settings

* Update vectorstore.py

* Update vectorstore.py

modify the comments of the param custom_index_settings

* Update vectorstore.py

* add settings conflict test

* reformat

---------

Co-authored-by: Quentin Pradet <quentin.pradet@elastic.co>
Co-authored-by: Miguel Grinberg <miguel.grinberg@gmail.com>