[docs] Add param notes to DSL search.delete documentation, fix broken links (#2861)
* Add param notes to DSL search.delete documentation, fix broken links

The preferred way of passing parameters to the DSL `Search.delete` appears to be calling `Search.params` first,
but this is only ever discussed in GitHub issues like
elastic/elasticsearch-dsl-py#1115
and elastic/elasticsearch-dsl-py#1395.

To help anyone else who has stumbled across this, I added a note about this to the documentation.

I also went ahead and updated the links for the `Search.scan` and `FacetedSearch.params`
methods to the most recent versions since they were 404ing.

* make sync and async changes the same

---------

Co-authored-by: Luna Lucadou <LUNA.LUCADOU@users.noreply.git.doit.wisc.edu>
Co-authored-by: Miguel Grinberg <miguel.grinberg@gmail.com>